+++ products/statement-generator/src/App.tsx
@@ -13,20 +13,20 @@ import AppHeader from 'page-layout/AppHeader';
 import AppSubheader from 'page-layout/AppSubheader';
 import Affirmation from 'page-layout/Affirmation';
 
-import BeforeYouBegin from 'flows/BeforeYouBegin';
-import IntroductionStep from 'flows/IntroductionStep';
-import InvolvementCommunityServiceFlow from 'involvement-step/InvolvementCommunityServiceFlow';
-import InvolvementInitialFlow from 'involvement-step/InvolvementInitialFlow';
-import InvolvementJobFlow from 'involvement-step/InvolvementJobFlow';
-import InvolvementParentingFlow from 'involvement-step/InvolvementParentingFlow';
-import InvolvementRecoveryFlow from 'involvement-step/InvolvementRecoveryFlow';
-import InvolvementSchoolFlow from 'involvement-step/InvolvementSchoolFlow';
-import InvolvementUnemployedFlow from 'involvement-step/InvolvementUnemployedFlow';
-import FinalizeForm from 'flows/FinalizeForm';
-import FinalizePreview from 'flows/FinalizePreview';
-import GoalsStep from 'flows/GoalsStep';
-import WhyStep from 'flows/WhyStep';
-import Download from 'flows/Download';
+import BeforeYouBegin from 'pages-form/BeforeYouBegin';
+import IntroductionStep from 'pages-form/IntroductionStep';
+import InvolvementCommunityServiceFlow from 'pages-form/InvolvementCommunityServiceFlow';
+import InvolvementInitialFlow from 'pages-form/InvolvementInitialFlow';
+import InvolvementJobFlow from 'pages-form/InvolvementJobFlow';
+import InvolvementParentingFlow from 'pages-form/InvolvementParentingFlow';
+import InvolvementRecoveryFlow from 'pages-form/InvolvementRecoveryFlow';
+import InvolvementSchoolFlow from 'pages-form/InvolvementSchoolFlow';
+import InvolvementUnemployedFlow from 'pages-form/InvolvementUnemployedFlow';
+import FinalizeForm from 'pages-form/FinalizeForm';
+import FinalizePreview from 'pages-form/FinalizePreview';
+import GoalsStep from 'pages-form/GoalsStep';
+import WhyStep from 'pages-form/WhyStep';
+import Download from 'pages-form/Download';
 
 import Landing from 'pages/Landing';
 import StartPage from 'pages/StartPage';
